fix: 헤더 컴포넌트에서 테마 키 호출 방식 수정
